requirements.txt
spidev
RPi.GPIO
rpi-lgpio
Pillow
fastapi
uvicorn
